audioreach-engine: Fix integer overflow in PCM MFC
capi_pcm_mf_cnv Guard against negative channel counts when
calculating required size for output media format.

pcm_mf_cnv_utils Use int32_t for loop counter to prevent overflow.

pc_init Use 64-bit constants in KPPS math to prevent unsigned
underflow.

[Ported to audioreach-engine-auto.cmn.1.0]
CRs-Fixed: 4491221

+++ modules/cmn/pcm_mf_cnv/capi/pcm_cnv/src/capi_pcm_mf_cnv_utils.cpp
@@ -415,7 +415,7 @@ static void *capi_pcm_mf_cnv_chmixer_get_coef_set_ptr(capi_pcm_mf_cnv_t *const m
    capi_pcm_mf_cnv_chmixer_coef_set_t *coef_set_ptr = me_ptr->config.coef_sets_ptr;
    int16_t *                           coef_ptr     = NULL;
 
-   for (int8_t i = me_ptr->config.num_coef_sets - 1; i >= 0; i--)
+   for (int32_t i = (int32_t) (me_ptr->config.num_coef_sets) - 1; i >= 0; i--)
    {
       if ((me_ptr->config.num_coef_sets - 1) != i) // except the first iteration
       {
